feat(#2101): Add javadocs for all new classes
