+++ eo-maven-plugin/src/test/java/org/eolang/maven/log/Logs.java
@@ -1,3 +1,26 @@
+/*
+ * The MIT License (MIT)
+ *
+ * Copyright (c) 2016-2023 Objectionary.com
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included
+ * in all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NON-INFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
 package org.eolang.maven.log;
 
 import java.util.Collection;
@@ -9,10 +32,35 @@ import java.util.concurrent.Executors;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
 
+/**
+ * Captured logs.
+ */
 public class Logs {
 
-    private final Collection<String> captured = new ConcurrentLinkedQueue<>();
+    /**
+     * Captured logs.
+     */
+    private final Collection<String> captured;
 
+    /**
+     * Ctor.
+     */
+    Logs() {
+        this(new ConcurrentLinkedQueue<>());
+    }
+
+    /**
+     * Ctor.
+     * @param captured Captured logs.
+     */
+    private Logs(final Collection<String> captured) {
+        this.captured = captured;
+    }
+
+    /**
+     * Return all captured logs up to this point.
+     * @return
+     */
     public Collection<String> captured() {
         return Collections.unmodifiableCollection(this.captured);
     }
@@ -62,6 +110,10 @@ public class Logs {
         }
     }
 
+    /**
+     * Append log message.
+     * @param log Log message.
+     */
     void append(final String log) {
         this.captured.add(log);
     }
